refactor: :label: move interfaces to external files
