+++ src/components/CustomHeader/CustomHeader.tsx
@@ -11,11 +11,8 @@ import {
 } from '@mantine/core'
 import { useBooleanToggle } from '@mantine/hooks'
 import { ColorSchemeToggle } from '../ColorSchemeToggle/ColorSchemeToggle'
+import { ICustomHeaderProps } from '../../ts/interfaces/CustomHeader.interface'
 import useStyles, { HEADER_HEIGHT } from './CustomHeader.styles'
-
-interface ICustomHeaderProps {
-  links: { link: string; label: string }[]
-}
 
 const CustomHeader = ({ links }: ICustomHeaderProps) => {
   const [opened, toggleOpened] = useBooleanToggle(false)
