+++ src/ts/interfaces/Article.interface.ts
@@ -1,0 +1,6 @@
+export interface IArticleProps {
+  thumbnail: string
+  title: string
+  link: string
+  categories: [string, string, string, string, string]
+}
