+++ src/ts/interfaces/CustomHeader.interface.ts
@@ -1,0 +1,3 @@
+export interface ICustomHeaderProps {
+  links: { link: string; label: string }[]
+}
