+++ src/ts/interfaces/CustomImage.interface.ts
@@ -1,0 +1,9 @@
+export interface ICustomImageProps {
+  src: string
+  alt: string
+  width?: number
+  height?: number
+  layout?: 'fixed' | 'fill' | 'intrinsic' | 'responsive' | 'raw'
+  objectFit?: any
+  className?: string
+}
