+++ src/ts/interfaces/Layout.interface.ts
@@ -1,0 +1,5 @@
+import { ReactNode } from 'react'
+
+export interface ILayoutProps {
+  children: ReactNode
+}
