+++ src/ts/interfaces/NowPlaying.type.ts
@@ -1,0 +1,8 @@
+export type NowPlayingSong = {
+  album: string
+  albumImageUrl: string
+  artist: string
+  isPlaying: boolean
+  songUrl: string
+  title: string
+}
